barebones_js/requirements.txt
Flask==2.2.5
Flask_Cors==4.0.0
flask_sqlalchemy==3.1.1
ipython==8.12.3
Jinja2==3.1.2
pandas==2.0.3
Pygments==2.14.0
PyYAML==6.0.2
Requests==2.32.3